+++ src/app/ngrx/actions/castle.action.creator.ts
@@ -1,7 +1,6 @@
 import { Action } from '@ngrx/store';
 import { ActionTypes } from './action.types';
-import { FilterModel, EpisodeModel, MetricsType } from '../model';
-import { PodcastModel } from '../reducers';
+import { FilterModel, EpisodeModel, PodcastModel, MetricsType } from '../';
 
 export interface CastleFilterPayload {
   filter: FilterModel;
