+++ src/app/ngrx/index.ts
@@ -1,3 +1,7 @@
-export * from './actions';
-export * from './model';
-export * from './reducers';
+export { MetricsType, getMetricsProperty } from './reducers/metrics.type';
+export { PodcastModel } from './reducers/podcast.reducer';
+export { EpisodeModel, EPISODE_PAGE_SIZE } from './reducers/episode.reducer';
+export { INTERVAL_MONTHLY, INTERVAL_WEEKLY, INTERVAL_DAILY, INTERVAL_HOURLY,
+  IntervalModel, IntervalList, FilterModel } from './reducers/filter.reducer';
+export { PodcastMetricsModel } from './reducers/podcast-metrics.reducer';
+export { EpisodeMetricsModel } from './reducers/episode-metrics.reducer';
