--- src/app/ngrx/model/filter.model.ts
@@ -1,12 +1,0 @@
-export const TODAY = 'Today';
-export const THIS_WEEK = 'This week';
-export const TWO_WEEKS = '2 weeks';
-export const THIS_MONTH = 'This month';
-export const THREE_MONTHS = '3 months';
-export const THIS_YEAR = 'This year';
-export const YESTERDAY = 'Yesterday';
-export const LAST_WEEK = 'Last week';
-export const PRIOR_TWO_WEEKS = 'Prior 2 weeks';
-export const LAST_MONTH = 'Last month';
-export const PRIOR_THREE_MONTHS = 'Prior 3 months';
-export const LAST_YEAR = 'Last year';
