--- src/app/ngrx/model/index.ts
@@ -1,5 +1,0 @@
-export { EpisodeModel } from '../reducers/episode.reducer';
-export { INTERVAL_MONTHLY, INTERVAL_WEEKLY, INTERVAL_DAILY, INTERVAL_HOURLY, IntervalModel, IntervalList, FilterModel } from '../reducers/filter.reducer';
-export { PodcastMetricsModel, EpisodeMetricsModel, MetricsType } from './metrics.model';
-export { TODAY, THIS_WEEK, TWO_WEEKS, THIS_MONTH, THREE_MONTHS, THIS_YEAR,
-  YESTERDAY, LAST_WEEK, PRIOR_TWO_WEEKS, LAST_MONTH, PRIOR_THREE_MONTHS, LAST_YEAR } from './filter.model';
