+++ src/app/shared/filter/date/standard-date-range.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, Output, EventEmitter, OnChanges } from '@angular/core';
-import { IntervalModel, INTERVAL_HOURLY, INTERVAL_DAILY, INTERVAL_WEEKLY, INTERVAL_MONTHLY,
+import { IntervalModel, INTERVAL_HOURLY, INTERVAL_DAILY, INTERVAL_WEEKLY, INTERVAL_MONTHLY } from '../../../ngrx';
+import { isMoreThanXDays, endOfTodayUTC,
+  beginningOfThreeMonthsUTC, beginningOfThisYearUTC,
   TODAY, THIS_WEEK, TWO_WEEKS, THIS_MONTH, THREE_MONTHS, THIS_YEAR,
-  YESTERDAY, LAST_WEEK, PRIOR_TWO_WEEKS, LAST_MONTH, PRIOR_THREE_MONTHS, LAST_YEAR } from '../../../ngrx/model';
-import { isMoreThanXDays, endOfTodayUTC, beginningOfTwoWeeksUTC, beginningOfThisMonthUTC,
-  beginningOfThreeMonthsUTC, beginningOfThisYearUTC } from '../../util/date.util';
+  YESTERDAY, LAST_WEEK, PRIOR_TWO_WEEKS, LAST_MONTH, PRIOR_THREE_MONTHS, LAST_YEAR } from '../../util/date.util';
 
 @Component({
   selector: 'metrics-standard-date-range',
